Introduce admin and dashboard and scope settings to admin

+++ db/seeds.rb
@@ -1,16 +1,7 @@
-# This file should contain all the record creation needed to seed the database with its default values.
-# The data can then be loaded with the rails db:seed command (or created alongside the database with db:setup).
-#
-# Examples:
-#
-#   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
-#   Character.create(name: 'Luke', movie: movies.first)
+if Rails.env == 'development'
+	User.destroy_all
+	AdminUser.destroy_all
 
-User.destroy_all
-
-User.create(
-				name: 'Some name',
-				email: 'user@test.com',
-				password: 'secret',
-				password_confirmation: 'secret'
-)
+	User.create( name: 'Some name', email: 'user@test.com', password: 'secret', password_confirmation: 'secret' )
+	AdminUser.create!(email: 'admin@test.com', password: 'secret', password_confirmation: 'secret')
+end
